requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
marshmallow==3.21.1
marshmallow-sqlalchemy==1.2.0
python-dotenv==1.0.1
